compositing: Split out `Painter` from `IOCompositor`
This change splits `IOCompositor` into two structs: `IOCompositor` and
`Painter`. The idea is that `Painter` will be per-`RenderingContext`.
Currently there is only one, but when there is more than one there will
be one `Painter` per-`RenderingContext`. Therefore, everything that we
suspect will be managed this way is moved to `Painter`.

In addition, to avoid the risk of double-borrows, `WebViewRender` no
longer keeps a reference to any of the containing structs. Instead
necessary things are passed down when necessary or stored internally.
