+++ components/compositing/lib.rs
@@ -24,7 +24,9 @@ pub use crate::compositor::{IOCompositor, WebRenderDebugOption};
 mod tracing;
 
 mod compositor;
+mod painter;
 mod pinch_zoom;
+mod pipeline_details;
 mod refresh_driver;
 mod render_notifier;
 mod screenshot;
